team_challenge/pyproject.toml
[tool.poetry]
name = "team-challenge"
version = "0.1.0"
description = "Dependencies for backend"
authors = ["0leksandrr <tsybulskyy.sasha@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django-import-export = "^3.3.7"
djoser = "^2.2.2"
celery = {extras = ["redis"], version = "^5.3.6"}
flower = "^2.0.1"
sqlparse = "^0.5.0"
gunicorn = "^22.0.0"
requests = "^2.32.2"
social-auth-app-django = "^5.4.1"
idna = "^3.7"
pillow = "^10.3.0"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.5.2"
pip = "^24.0"
install = "^1.3.5"
factory-boy = "^3.3.0"



[tool.poetry.group.dev.dependencies]
asgiref = "3.7.2"
cachetools = "5.3.2"
certifi = "2024.2.2"
charset-normalizer = "3.3.2"
django = "4.2.11"
django-cors-headers = "4.3.1"
django-extensions = "3.2.3"
django-filter = "24.1"
django-storages = "1.14.2"
djangorestframework = "3.14.0"
drf-yasg = "1.21.7"
google-api-core = "2.17.0"
google-auth = "2.27.0"
google-cloud-core = "2.4.1"
google-cloud-storage = "2.14.0"
google-crc32c = "1.5.0"
google-resumable-media = "2.7.0"
googleapis-common-protos = "1.62.0"
inflection = "0.5.1"
packaging = "23.2"
protobuf = "4.25.2"
psycopg2-binary = "2.9.9"
pyasn1 = "0.5.1"
pyasn1-modules = "0.3.0"
python-dotenv = "1.0.1"
pytz = "2023.3.post1"
pyyaml = "6.0.1"
rsa = "4.9"
setuptools = "69.0.3"
tzdata = "2023.4"
uritemplate = "4.1.1"
urllib3 = "2.2.1"
waitress = "3.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
